refactor: extract EditTitleModal into standalone file
